requirements.txt
wxPython==4.0.0b2
PyOpenGL==3.1.0
bvh==0.3
